assembler: remove original assembler
