--- assembler/src/ir/ir4_parse_ambiguous_tokens.rs
@@ -1,520 +1,0 @@
-use std::convert::TryInto;
-use num_traits::Num;
-use std::string::ToString;
-use lc3_isa::{Addr, SignedWord, check_signed_imm, Word};
-
-use crate::error::{ParseError, InvalidLabelReason, InvalidRegReason, InvalidImmediateReason, Error};
-use crate::lexer::Token;
-use crate::ir::ir2_parse_line_syntax;
-use crate::ir::ir3_parse_objects;
-use crate::parser::LeniencyLevel;
-
-#[derive(Clone, Debug)]
-pub struct File<'input> {
-    pub objects: Vec<Object<'input>>,
-    pub ignored: Vec<ir2_parse_line_syntax::Line<'input>>,
-}
-
-#[derive(Clone, Debug)]
-pub struct Object<'input> {
-    pub origin_src: Operation<'input>,
-    pub origin: Immediate<'input, Addr>,
-    pub content: ObjectContent<'input>,
-}
-
-#[derive(Clone, Debug)]
-pub struct ObjectContent<'input> {
-    pub operations: Vec<Operation<'input>>,
-    pub empty_lines: Vec<ir2_parse_line_syntax::Line<'input>>,
-    pub hanging_labels: Vec<ir2_parse_line_syntax::Line<'input>>,
-    pub invalid_lines: Vec<ir2_parse_line_syntax::Line<'input>>,
-}
-
-pub type Label<'input> = Checked<'input, &'input str>;
-pub type Separator<'input> = Token<'input>;
-
-// Different from lc3_isa::Instruction in that offsets from labels aren't computed.
-// Also covers pseudo-ops.
-#[derive(Clone, Debug)]
-pub struct Operation<'input> {
-    pub label: Option<Label<'input>>,
-    pub operator: Token<'input>,
-    pub nzp: Result<Option<ConditionCodes>, ParseError>,
-    pub operands: Operands<'input>,
-
-    pub src_lines: Vec<String>,
-    pub separators: Vec<Separator<'input>>,
-    pub whitespace: Vec<Token<'input>>,
-    pub comments: Vec<Token<'input>>,
-    pub newlines: Vec<Token<'input>>,
-}
-
-#[derive(Clone, Debug)]
-pub struct Checked<'input, T> {
-    pub src: Token<'input>,
-    pub value: Result<T, ParseError>,
-}
-
-impl<'input, T> Checked<'input, T> {
-    pub fn unwrap(self) -> T {
-        self.value.unwrap()
-    }
-
-    pub fn extract_error_into(&self, errors: &mut Vec<Error>) {
-        if let Err(error) = &self.value {
-            errors.push(error.into());
-        }
-    }
-}
-
-pub type Reg<'input> = Checked<'input, lc3_isa::Reg>;
-pub type Immediate<'input, T> = Checked<'input, T>;
-
-#[derive(Clone, Debug)]
-pub enum Sr2OrImm5<'input> {
-    Sr2(Reg<'input>),
-    Imm5(Immediate<'input, SignedWord>),
-}
-
-#[derive(Clone, Debug)]
-pub enum ImmOrLabel<'input> {
-    Imm(Immediate<'input, SignedWord>),
-    Label(Label<'input>),
-}
-
-#[derive(Clone, Debug)]
-pub enum UnsignedImmOrLabel<'input> {
-    Imm(Immediate<'input, Word>),
-    Label(Label<'input>),
-}
-
-#[derive(Clone, Debug)]
-pub struct ConditionCodes {
-    pub n: bool,
-    pub z: bool,
-    pub p: bool,
-}
-
-type PCOffset<'input> = Checked<'input, ImmOrLabel<'input>>;
-
-#[derive(Clone, Debug)]
-pub enum Operands<'input> {
-    Add { dr: Reg<'input>, sr1: Reg<'input>, sr2_or_imm5: Checked<'input, Sr2OrImm5<'input>> },
-    And { dr: Reg<'input>, sr1: Reg<'input>, sr2_or_imm5: Checked<'input, Sr2OrImm5<'input>> },
-    Br { pc_offset9: PCOffset<'input> },
-    Jmp { base: Reg<'input> },
-    Jsr { pc_offset11: PCOffset<'input> },
-    Jsrr { base: Reg<'input> },
-    Ld { dr: Reg<'input>, pc_offset9: PCOffset<'input> },
-    Ldi { dr: Reg<'input>, pc_offset9: PCOffset<'input> },
-    Ldr { dr: Reg<'input>, base: Reg<'input>, offset6: Immediate<'input, SignedWord> },
-    Lea { dr: Reg<'input>, pc_offset9: PCOffset<'input> },
-    Not { dr: Reg<'input>, sr: Reg<'input> },
-    Ret,
-    Rti,
-    St { sr: Reg<'input>, pc_offset9: PCOffset<'input> },
-    Sti { sr: Reg<'input>, pc_offset9: PCOffset<'input> },
-    Str { sr: Reg<'input>, base: Reg<'input>, offset6: Immediate<'input, SignedWord> },
-    Trap { trap_vec: Immediate<'input, u8> },
-
-    Getc,
-    Out,
-    Puts,
-    In,
-    Putsp,
-    Halt,
-
-    Orig { origin: Immediate<'input, Addr> },
-    Fill { value: Checked<'input, UnsignedImmOrLabel<'input>> },
-    Blkw { size_src: Token<'input>, size: Immediate<'input, Addr> }, // Addr used here to signify a number of locations. Max is number of possible Addrs.
-    Stringz { string: Checked<'input, String> },
-    End,
-}
-
-pub struct AmbiguousTokenParser {
-    pub leniency: LeniencyLevel,
-}
-
-impl AmbiguousTokenParser {
-
-    pub fn parse_ambiguous_tokens<'input>(&self, file: ir3_parse_objects::File<'input>) -> File<'input> {
-        let ir3_parse_objects::File { objects, ignored } = file;
-        File {
-            objects: objects.into_iter().map(|o| self.validate_object(o)).collect(),
-            ignored
-        }
-    }
-
-    fn validate_object<'input>(&self, object: ir3_parse_objects::Object<'input>) -> Object<'input> {
-        let ir3_parse_objects::Object { origin_src, origin, content } = object;
-        let ir3_parse_objects::ObjectContent { operations, empty_lines, hanging_labels, invalid_lines } = content;
-        Object {
-            origin_src: self.validate_line(origin_src),
-            origin: self.validate_numeric_immediate(origin),
-            content: ObjectContent {
-                operations: operations.into_iter().map(|o| self.validate_line(o)).collect(),
-                empty_lines,
-                hanging_labels,
-                invalid_lines
-            }
-        }
-    }
-
-    fn validate_line<'input>(&self, line: ir3_parse_objects::Line<'input>) -> Operation<'input> {
-        let ir3_parse_objects::Line {
-            label,
-            operation: ir2_parse_line_syntax::OperationTokens {
-                operator,
-                operands,
-                separators,
-            },
-            whitespace,
-            comments,
-            newlines,
-            src_lines,
-        } = line.clone();
-
-        Operation {
-            label: label.map(|l| self.validate_label(l)),
-            operator,
-            nzp: self.validate_condition_codes(&operator),
-            operands: self.validate_operand_tokens(operands),
-            separators,
-            whitespace,
-            comments,
-            newlines,
-            src_lines,
-        }
-    }
-
-    fn validate_operand_tokens<'input>(&self, operands: ir2_parse_line_syntax::OperandTokens<'input>) -> Operands<'input> {
-        use ir2_parse_line_syntax::OperandTokens;
-        match operands {
-            OperandTokens::Add { dr, sr1, sr2_or_imm5 } =>
-                Operands::Add {
-                    dr: self.validate_reg(dr),
-                    sr1: self.validate_reg(sr1),
-                    sr2_or_imm5: self.validate_sr2_or_imm5(sr2_or_imm5)
-                },
-            OperandTokens::And { dr, sr1, sr2_or_imm5 } =>
-                Operands::And {
-                    dr: self.validate_reg(dr),
-                    sr1: self.validate_reg(sr1),
-                    sr2_or_imm5: self.validate_sr2_or_imm5(sr2_or_imm5)
-                },
-            OperandTokens::Br { label } => {
-                Operands::Br { pc_offset9: self.validate_imm_or_label(label, 9), }
-            },
-            OperandTokens::Jmp { base } => Operands::Jmp { base: self.validate_reg(base) },
-            OperandTokens::Jsr { label } => Operands::Jsr { pc_offset11: self.validate_imm_or_label(label, 11) },
-            OperandTokens::Jsrr { base } => Operands::Jsrr { base: self.validate_reg(base) },
-            OperandTokens::Ld { dr, label } => Operands::Ld { dr: self.validate_reg(dr), pc_offset9: self.validate_imm_or_label(label, 9) },
-            OperandTokens::Ldi { dr, label } => Operands::Ldi { dr: self.validate_reg(dr), pc_offset9: self.validate_imm_or_label(label, 9) },
-            OperandTokens::Ldr { dr, base, offset6 } =>
-                Operands::Ldr {
-                    dr: self.validate_reg(dr),
-                    base: self.validate_reg(base),
-                    offset6: self.validate_signed_immediate(offset6, 6),
-                },
-            OperandTokens::Lea { dr, label } => Operands::Lea { dr: self.validate_reg(dr), pc_offset9: self.validate_imm_or_label(label, 9) },
-            OperandTokens::Not { dr, sr } => Operands::Not { dr: self.validate_reg(dr), sr: self.validate_reg(sr) },
-            OperandTokens::Ret => Operands::Ret,
-            OperandTokens::Rti => Operands::Rti,
-            OperandTokens::St { sr, label } => Operands::St { sr: self.validate_reg(sr), pc_offset9: self.validate_imm_or_label(label, 9) },
-            OperandTokens::Sti { sr, label } => Operands::Sti { sr: self.validate_reg(sr), pc_offset9: self.validate_imm_or_label(label, 9) },
-            OperandTokens::Str { sr, base, offset6 } =>
-                Operands::Str {
-                    sr: self.validate_reg(sr),
-                    base: self.validate_reg(base),
-                    offset6: self.validate_signed_immediate(offset6, 6),
-                },
-            OperandTokens::Trap { trap_vec } => Operands::Trap { trap_vec: self.validate_numeric_immediate(trap_vec) },
-
-            OperandTokens::Getc => Operands::Getc,
-            OperandTokens::Out => Operands::Out,
-            OperandTokens::Puts => Operands::Puts,
-            OperandTokens::In => Operands::In,
-            OperandTokens::Putsp => Operands::Putsp,
-            OperandTokens::Halt => Operands::Halt,
-
-            OperandTokens::Orig { origin } => Operands::Orig { origin: self.validate_numeric_immediate(origin) },
-            OperandTokens::Fill { value } => Operands::Fill { value: self.validate_unsigned_imm_or_label(value) },
-            OperandTokens::Blkw { size } => Operands::Blkw { size_src: size, size: self.validate_blkw_immediate(size) },
-            OperandTokens::Stringz { string } => Operands::Stringz { string: self.validate_string(string) },
-            OperandTokens::End => Operands::End,
-        }
-    }
-
-    fn validate_sr2_or_imm5<'input>(&self, src: Token<'input>) -> Checked<'input, Sr2OrImm5<'input>> {
-        let reg = self.validate_reg(src);
-        let imm5 = self.validate_signed_immediate(src, 5);
-        let value = if let Reg { value: Ok(_), .. } = reg {
-            Ok(Sr2OrImm5::Sr2(reg))
-        } else if let Immediate { value: Ok(_), .. } = imm5 {
-            Ok(Sr2OrImm5::Imm5(imm5))
-        } else {
-            if let Reg { value: Err(ParseError::InvalidReg { reason: invalid_reg_reason, .. }), .. } = reg {
-                if let Immediate { value: Err(ParseError::InvalidImmediate { reason: invalid_imm5_reason, .. }), .. } = imm5 {
-                    Err(ParseError::InvalidRegOrImm5 {
-                        range: src.span,
-                        invalid_reg_reason,
-                        invalid_imm5_reason,
-                    })
-                } else {
-                    unreachable!()
-                }
-            } else {
-                unreachable!() // TODO: use something cleaner like a match for this
-            }
-        };
-        Checked { src, value }
-    }
-
-    fn validate_reg<'input>(&self, src: Token<'input>) -> Reg<'input> {
-        let value = if let Some("r") | Some("R") = src.src.get(..=0) {
-            src.src.get(1..)
-                .filter(|s| s.len() == 1)
-                .and_then(|s| s.parse::<u8>().ok())
-                .and_then(|i| i.try_into().ok())
-                .ok_or(ParseError::InvalidReg {
-                    range: src.span,
-                    reason: InvalidRegReason::Number,
-                })
-        } else {
-            Err(ParseError::InvalidReg {
-                range: src.span,
-                reason: InvalidRegReason::FirstChar,
-            })
-        };
-        Reg { src, value }
-    }
-
-    fn validate_numeric_immediate<'input, T: Num>(&self, src: Token<'input>) -> Immediate<'input, T> {
-        let Token { src: str, span, .. } = src;
-        let value = if let Some(str_head) = str.get(..=0) {
-            let (str_head, offset) = match str.get(0..2) {
-                Some("0b") | Some("0x") => (str.get(1..2).unwrap(), 2),
-                Some(_) => (str_head, 1),
-                // If we don't have two chars, just pass it along?
-                None => (str_head, 1),
-            };
-
-            let radix = match str_head {
-                "b" => Some(2),
-                "#" => Some(10),
-                "x" => Some(16),
-                _ => None
-            };
-            if let Some(radix) = radix {
-                if let Some(src_tail) = src.src.get(offset..) {
-                    T::from_str_radix(src_tail, radix)
-                        .map_err(|_| InvalidImmediateReason::Number { actual: src_tail.to_string() })
-                } else {
-                    Err(InvalidImmediateReason::NoNumber)
-                }
-            } else {
-                Err(InvalidImmediateReason::RadixChar { actual: str_head.to_string() })
-            }
-        } else {
-            Err(InvalidImmediateReason::NoChars)
-        }.map_err(|reason| ParseError::InvalidImmediate {
-            range: span,
-            reason
-        });
-
-        Immediate { src, value }
-    }
-
-    fn validate_signed_immediate<'input>(&self, src: Token<'input>, num_bits: u32) -> Immediate<'input, SignedWord> {
-        let Immediate { src, value } = self.validate_numeric_immediate(src);
-        let value = match value {
-            Ok(i) => {
-                if check_signed_imm(i, num_bits) {
-                    Ok(i)
-                } else {
-                    Err(ParseError::InvalidImmediate {
-                        range: src.span,
-                        reason: InvalidImmediateReason::OutOfRange { value: i, num_bits }
-                    })
-                }
-            }
-            error => error // TODO: look for appropriate combinator(s)?
-        };
-        Immediate { src, value }
-    }
-
-    fn validate_imm_or_label<'input>(&self, src: Token<'input>, num_bits: u32) -> Checked<'input, ImmOrLabel<'input>>  {
-        let label = self.validate_label(src);
-        let imm = self.validate_signed_immediate(src, num_bits);
-        let value = if let Label { value: Ok(_), .. } = label {
-            Ok(ImmOrLabel::Label(label))
-        } else if let Immediate { value: Ok(_), .. } = imm {
-            Ok(ImmOrLabel::Imm(imm))
-        } else {
-            if let Label { value: Err(ParseError::InvalidLabel { reasons: invalid_label_reasons, .. }), .. } = label {
-                if let Immediate { value: Err(ParseError::InvalidImmediate { reason: invalid_immediate_reason, .. }), .. } = imm {
-                    Err(ParseError::InvalidLabelOrImmediate {
-                        range: src.span,
-                        invalid_label_reasons,
-                        invalid_immediate_reason
-                    })
-                } else {
-                    unreachable!()
-                }
-            } else {
-                unreachable!() // TODO: use something cleaner like a match for this
-            }
-        };
-        Checked { src, value }
-    }
-
-    fn validate_unsigned_imm_or_label<'input>(&self, src: Token<'input>) -> Checked<'input, UnsignedImmOrLabel<'input>>  {
-        let label = self.validate_label(src);
-        let imm = self.validate_numeric_immediate(src);
-        let value = if let Immediate { value: Ok(_), .. } = imm {
-            Ok(UnsignedImmOrLabel::Imm(imm))
-        } else if let Label { value: Ok(_), .. } = label {
-            Ok(UnsignedImmOrLabel::Label(label))
-        } else {
-            if let Label { value: Err(ParseError::InvalidLabel { reasons: invalid_label_reasons, .. }), .. } = label {
-                if let Immediate { value: Err(ParseError::InvalidImmediate { reason: invalid_immediate_reason, .. }), .. } = imm {
-                    Err(ParseError::InvalidLabelOrImmediate {
-                        range: src.span,
-                        invalid_label_reasons,
-                        invalid_immediate_reason
-                    })
-                } else {
-                    unreachable!()
-                }
-            } else {
-                unreachable!() // TODO: use something cleaner like a match for this
-            }
-        };
-        Checked { src, value }
-    }
-
-    fn validate_label<'input>(&self, src: Token<'input>) -> Label<'input> {
-        let label = src.src;
-
-        let length = label.len();
-        let valid_length = if self.leniency.long_labels_allowed() {
-            length >= 1
-        } else {
-            (1..=20).contains(&length)
-        };
-
-        let mut chars = label.chars();
-        let first_char = chars.next();
-        let first_char_alphabetic = first_char.filter(|c| c.is_alphabetic()).is_some();
-
-        let mut other_chars = chars.collect::<Vec<_>>();
-        other_chars.retain(|&c| !(c.is_alphanumeric() || c == '_'));
-        let other_chars_alphanumeric = other_chars.len() == 0;
-
-        let mut invalidation_reasons = Vec::new();
-        if !valid_length {
-            invalidation_reasons.push(InvalidLabelReason::Length { actual: length.clone() });
-        }
-        if !first_char_alphabetic {
-            invalidation_reasons.push(InvalidLabelReason::FirstChar { actual: first_char });
-        }
-        if !other_chars_alphanumeric {
-            invalidation_reasons.push(InvalidLabelReason::OtherChars { actual: other_chars.into_iter().collect::<String>() });
-        }
-
-        let value = if invalidation_reasons.len() == 0 {
-            Ok(label)
-        } else {
-            Err(ParseError::InvalidLabel {
-                range: src.span,
-                reasons: invalidation_reasons,
-            })
-        };
-
-        Label { src, value }
-    }
-
-    fn validate_condition_codes(&self, src: &Token) -> Result<Option<ConditionCodes>, ParseError> {
-        let str = src.src;
-        if str.to_uppercase().starts_with("BR") {
-            let mut n = false;
-            let mut z = false;
-            let mut p = false;
-            for c in str[2..].to_lowercase().chars() {
-                match c {
-                    // TODO: prettify with macro or non-iterative solution
-                    'n' => {
-                        if n { return Err(ParseError::Misc("Duplicate condition code n.".to_string())); }
-                        n = true;
-                    },
-                    'z' => {
-                        if z { return Err(ParseError::Misc("Duplicate condition code z.".to_string())); }
-                        z = true;
-                    },
-                    'p' => {
-                        if p { return Err(ParseError::Misc("Duplicate condition code p.".to_string())); }
-                        p = true;
-                    },
-                    _ => { return Err(ParseError::Misc("Invalid condition codes.".to_string())) },
-                }
-            }
-            if !(n || z || p) {
-                n = true;
-                z = true;
-                p = true;
-            }
-            Ok(Some(ConditionCodes { n, z, p }))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn validate_blkw_immediate<'input>(&self, src: Token<'input>) -> Immediate<'input, Addr> {
-        Immediate {
-            src,
-            value: src.src.parse().map_err(|_| ParseError::Misc("Invalid BLKW immediate.".to_string()))
-        }
-    }
-
-    fn validate_string<'input>(&self, src: Token<'input>) -> Checked<'input, String> {
-        let mut string = src.src.to_string();
-        // remove start and end quote
-        string.pop();
-        string.remove(0);
-        // remove escape characters
-        string = string
-            .replace(r#"\""#, r#"""#)
-            .replace(r#"\\"#, r#"\"#) // TODO: fix this logic to escape \\ properly (atm \\n becomes '\n', not '\' and 'n')
-            .replace(r#"\n"#, "\n");
-        let value = Ok(string);
-        Checked { src, value }
-    }
-}
-
-#[cfg(test)]
-mod immediate_tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    fn single_test<N: core::fmt::Debug + Num>(num: &str, actual: N)  {
-        let p = AmbiguousTokenParser { leniency: LeniencyLevel::Lenient };
-
-        let tok = Token { src: num, span: (0, 0), ty: crate::lexer::TokenType::Ambiguous };
-
-        assert_eq!(actual, p.validate_numeric_immediate(tok).value.unwrap());
-    }
-
-    #[test]
-    fn regular() {
-        single_test("0x123", 0x123);
-        single_test("0x0123", 0x0123);
-        single_test("0b0101", 0b0101);
-    }
-
-    #[test]
-    fn patt_style() {
-        single_test("#100", 100);
-        single_test("x456", 0x456);
-        single_test("b0101", 0b0101);
-    }
-}
